Move merger to broker
