+++ chronicle-cli/Cargo.toml
@@ -7,8 +7,7 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-backstage = "0.1"
-chronicle = { path = "../chronicle" }
+chronicle = { path = "../chronicle", default-features = false }
 chronicle-common = { path = "../chronicle-common" }
 chronicle-broker = { path = "../chronicle-broker", default-features = false }
 scylla-rs = "0.1"
@@ -22,5 +21,3 @@ serde_json = "1.0"
 anyhow = "1.0"
 regex = "1.4"
 indicatif = "0.16"
-glob = "0.3"
-thiserror = "1.0"
