+++ chronicle/Cargo.toml
@@ -14,6 +14,7 @@ path = "src/lib.rs"
 [[bin]]
 name = "chronicle"
 path = "src/main.rs"
+required-features = ["application"]
 
 [dependencies]
 backstage = "0.1"
@@ -21,7 +22,7 @@ scylla-rs = "0.1"
 chronicle-common = { path = "../chronicle-common" }
 chronicle-api = { path = "../chronicle-api" }
 chronicle-storage = { path = "../chronicle-storage" }
-chronicle-broker = { path = "../chronicle-broker", default-features = false, optional = true }
+chronicle-broker = { path = "../chronicle-broker", default-features = false }
 serde = { version = "1.0", features = ["derive"] }
 async-trait = "0.1"
 env_logger = "0.8"
@@ -37,7 +38,7 @@ anyhow = "1.0"
 tokio = "1.5"
 
 [features]
-default = ["rocket", "chronicle-broker"]
+default = ["rocket", "application"]
 application = ["chronicle-broker/application"]
 filter = ["chronicle-broker/filter"]
 rocket = ["chronicle-api/rocket_listener"]
